Modulized and updated readme
